Moved `Calibration` to separate module
